[Stream][ActionAdventure][Items][Create classes]
Add Items classes
