+++ ue5/ActionAdventure/Plugins/Developer/RiderLink/Intermediate/FileSystemMappings.ini
@@ -1,0 +1,7 @@
+/Engine=C:/Program Files (x86)/UE_5.3/Engine/Shaders
+/ShaderAutogen=E:/UE_Projects/UE5/ActionAdventure/ActionAdventure/ue5/ActionAdventure/Intermediate/ShaderAutogen
+/Plugin/FX/Niagara=C:/Program Files (x86)/UE_5.3/Engine/Plugins/FX/Niagara/Shaders
+/Plugin/GLTFExporter=C:/Program Files (x86)/UE_5.3/Engine/Plugins/Enterprise/GLTFExporter/Shaders
+/Plugin/ExrReaderShaders=C:/Program Files (x86)/UE_5.3/Engine/Plugins/Media/ImgMedia/Shaders
+/Plugin/WmfMedia=C:/Program Files (x86)/UE_5.3/Engine/Plugins/Media/WmfMedia/Shaders
+/Plugin/Experimental/ChaosNiagara=C:/Program Files (x86)/UE_5.3/Engine/Plugins/Experimental/ChaosNiagara/Shaders
