+++ ue5/ActionAdventure/Saved/Config/CrashReportClient/UECC-Windows-1330DD4E4D8B858B5C3073895B2516FB/CrashReportClient.ini
@@ -1,0 +1,7 @@
+[CrashReportClient]
+bHideLogFilesOption=false
+bIsAllowedToCloseWithoutSending=true
+CrashConfigPurgeDays=2
+Stall.RecordDump=false
+Ensure.RecordDump=true
+
