+++ ue5/ActionAdventure/Saved/Config/WindowsEditor/Editor.ini
@@ -1,0 +1,3 @@
+[/Script/UnrealEd.UnrealEdOptions]
+UsingXGE=True
+
