+++ ue5/ActionAdventure/Saved/Config/WindowsEditor/Engine.ini
@@ -46,14 +46,6 @@ Paths=../../../Engine/Plugins/Media/MediaPlate/Content
 Paths=../../../Engine/Plugins/Editor/ObjectMixer/ObjectMixer/Content
 Paths=../../../Engine/Plugins/Editor/ObjectMixer/LightMixer/Content
 
-[WindowsApplication.Accessibility]
-StickyKeysHotkey=False
-ToggleKeysHotkey=False
-FilterKeysHotkey=False
-StickyKeysConfirmation=False
-ToggleKeysConfirmation=False
-FilterKeysConfirmation=False
-
 [/Script/UdpMessaging.UdpMessagingSettings]
 EnabledByDefault=False
 EnableTransport=True
@@ -75,4 +67,8 @@ TunnelMulticastEndpoint=
 SDKPath=(Path="")
 NDKPath=(Path="")
 JavaPath=(Path="")
+
+[/Script/UnrealEd.UnrealEdEngine]
+TemplateMapInfos=(ThumbnailTexture=None,Thumbnail="/Engine/Maps/Templates/Thumbnails/OpenWorld.OpenWorld",Map="/Engine/Maps/Templates/OpenWorld",DisplayName="Open World",Category="OpenWorld")
+TemplateMapInfos=(ThumbnailTexture=None,Thumbnail="/Engine/Maps/Templates/Thumbnails/Basic.Basic",Map="/Engine/Maps/Templates/Template_Default",DisplayName="Basic",Category="")
