+++ ue5/ActionAdventure/Saved/Config/WindowsEditor/SourceControlSettings.ini
@@ -1,0 +1,3 @@
+[SourceControl.SourceControlSettings]
+Provider=Git
+
