--- ue5/ActionAdventure/Source/ActionAdventure/ActionAdventureGameMode.h
@@ -1,19 +1,0 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "GameFramework/GameModeBase.h"
-#include "ActionAdventureGameMode.generated.h"
-
-UCLASS(minimalapi)
-class AActionAdventureGameMode : public AGameModeBase
-{
-	GENERATED_BODY()
-
-public:
-	AActionAdventureGameMode();
-};
-
-
-
