+++ ue5/ActionAdventure/Source/ActionAdventure/GameCore/AAGameMode.h
@@ -1,0 +1,19 @@
+// Copyright Epic Games, Inc. All Rights Reserved.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/GameModeBase.h"
+#include "ActionAdventureGameMode.generated.h"
+
+UCLASS(minimalapi)
+class AActionAdventureGameMode : public AGameModeBase
+{
+	GENERATED_BODY()
+
+public:
+	AActionAdventureGameMode();
+};
+
+
+
